Create new followables helper to have some methods previously on follows helper. Rename entity to followable.
