+++ app/views/users/_following.html.erb
@@ -6,7 +6,7 @@
 
       <a href="#" class="accordion-title">
         <span class="icon">
-          <i class="icon icon-<%= entity_icon(followable_type) %>"></i>
+          <i class="icon icon-<%= followable_icon(followable_type) %>"></i>
         </span>
         <span class="title">
           <strong><%= entity_title(followable_type) %></strong>
