+++ config/locales/en/general.yml
@@ -504,7 +504,7 @@ en:
     flag: Flag as inappropriate
     follow: "Follow"
     follow_entity: "Follow %{entity}"
-    follow_entity_title: "Follow %{entity}: You can participate and receive notifications of any related events."
+    followable_title: "Follow %{entity}: You can participate and receive notifications of any related events."
     hide: Hide
     print:
       print_button: Print this info
